# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "cristinaperezboix.alu@iespacomolla.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/games/{id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Muestra un juego por ID",
                "description": "Muestra un juego por ID.",
                "operationId": "bd823a9a8a3dde40ae85d4481b6b1351",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del juego",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vista para mostrar el juego"
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            }
        },
        "/api/games": {
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Almacena un nuevo juego",
                "description": "Crea un nuevo registro de juego en la base de datos.",
                "operationId": "store",
                "responses": {
                    "302": {
                        "description": "Redirecciona a la página de creación de personajes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "game_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/games/{game_id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Muestra los detalles de un juego específico",
                "description": "Recupera los detalles de un juego específico en la base de datos.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "game_id",
                        "in": "path",
                        "description": "ID del juego",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve la vista con los detalles del juego"
                    },
                    "404": {
                        "description": "Juego no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Juego no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/games/finish": {
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Finaliza un juego",
                "description": "Actualiza el estado de un juego a 'Finalizado' en la base de datos.",
                "operationId": "finish",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "game_id"
                                ],
                                "properties": {
                                    "game_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirecciona al evento de finalización del juego",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "character_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Juego no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Juego no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/games/{id}": {
            "delete": {
                "tags": [
                    "Games"
                ],
                "summary": "Elimina un juego",
                "description": "Elimina un juego específico de la base de datos.",
                "operationId": "destroyGame",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del juego",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirecciona a la página de inicio del dashboard",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Successful game delete!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Juego no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Game not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos para eliminar el juego",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have no permission! You are not the owner of this game!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}